+++ CHANGES.md
@@ -22,6 +22,10 @@
 - **irmin-pack**
   - Allow snapshot export to work on indexed root nodes (#1845, @icristescu)
 
+- **irmin**
+  - Fix Tree.export for nodes exported twice using different repos. (#1795,
+    @Ngoguey42)
+
 ## 3.2.1 (2022-04-07)
 
 - Support all version of cmdliner (#1803, @samoht)